Use DBL instead of INT for MIN and MAX macros

+++ wayfire-config-manager.cpp
@@ -159,13 +159,13 @@ get_plugin_data(Plugin *p, Option *opt, xmlDoc *doc, xmlNode * a_node)
 					printf("%s\n", prop);
 					if (string((char *) prop) == "int") {
 						o->type = OPTION_TYPE_INT;
-						o->data.min = INT_MIN;
-						o->data.max = INT_MAX;
+						o->data.min = DBL_MIN;
+						o->data.max = DBL_MAX;
 						o->data.precision = 0.1;
 					} else if (string((char *) prop) == "double") {
 						o->type = OPTION_TYPE_DOUBLE;
-						o->data.min = INT_MIN;
-						o->data.max = INT_MAX;
+						o->data.min = DBL_MIN;
+						o->data.max = DBL_MAX;
 						o->data.precision = 0.1;
 					} else if (string((char *) prop) == "bool") {
 						o->type = OPTION_TYPE_BOOL;
